Added a removeRole function, works the same way as approveRole.
Added an option for Owners to verify transactions as well with their unique role, instead of adding them a role of "Verifier".

Added an "Owner" role in management, the owner of the contract is added automatically to the `roles` mapping when the contract is deployed (using the constractor).

Updated structs' ID's to 64 bits instead of 256 (default) to save on gas executions.

Successfully implemented the deposit bottles functions; the depositer will request to deposit a certain amount of bottles, a verifier will approve it, and once it is approved - the depositer will receive 2 tokens per bottle.

Future:
1. Still no events declared, will be declared later.
2. Spend coins contract needs to be implemented as well.
3. Switch current management process with ownable.sol by OpenZeppelin.