Added name & balance to greeners array in depositor. Created a function that adds people to the array, instead of the requestDeposition function (since a new user will be added to the array every time a user requests a deposition).
Added a modifier to check if user is registered; so certain functions (such as requestDeposition) will not be available to unregistered users.

Added a shop contract for shops to register themselves as shops.

Added a modifier to EcoCoin for requiring certain functions only for owners (similar to Ownable).

Added a spender contract for recyclers to spend their tokens at shops.

Added a redeem contract for recyclers to redeem their tokens for real value; added a "Redeemer" role in manager contract, the only one that can approve redeeming tokens (not even owner). (not fully functional yet)

Updated the import of ERC20 from openzeppelin directly from directory, instead of GitHub, to prevent future problems of code changes.

Future:
1. Send automatically 0.002 ether to every user that signs up for the token, so he can pay for the gases (Max fee:  0.00061739 SepoliaETH).