Updated Municipality & EcoToken
Added library for implementing the municipality struct, and a contract within the same file to store the data of the municipality.

Main error as of now is the option to mint and burn tokens wihtout fully inherting on instancing the EcoToken,  since it requires the contstructor variables to be set each time.